fix(fretboard): preserve boundary precision

+++ app/src/main/java/com/a3322505a/guitarlearning/NoteNameTrainingScreen.kt
@@ -335,7 +335,7 @@ internal fun noteTrainingFretboardWidthFraction(range: NoteTrainingRange): Float
         firstFret = range.fretRange.first,
         lastFret = range.fretRange.last,
     )
-    return (visibleWidth / referenceWidth).coerceAtMost(1f)
+    return (visibleWidth / referenceWidth).coerceAtMost(1.0).toFloat()
 }
 
 private fun markerFor(
